requirements_prod.txt
accelerate==1.10.1
affine==2.4.0
aiofiles==23.2.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
attrs==25.3.0
av==15.1.0
braceexpand==0.1.7
Brotli==1.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
click-plugins==1.1.1.2
cligj==0.7.2
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
einops==0.8.1
executing==2.2.1
faiss-cpu==1.12.0
fastapi==0.118.0
ffmpy==0.6.1
filelock==3.19.1
fonttools==4.60.0
fsspec==2025.9.0
gradio==5.47.2
gradio_client==1.13.3
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.2
idna==3.10
importlib_resources==6.5.2
ipykernel==6.30.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
narwhals==2.6.0
nest-asyncio==1.6.0
networkx==3.5
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
orjson==3.11.3
packaging==25.0
pandas==2.3.2
parso==0.8.5
peft==0.13.2
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.9
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
qwen-vl-utils==0.0.14
rasterio==1.4.3
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
rich==14.1.0
rpds-py==0.27.1
ruff==0.13.2
safehttpx==0.1.6
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
semantic-version==2.10.0
sentence-transformers==3.0.1
sentencepiece==0.2.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.48.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
tomlkit==0.12.0
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.2
triton==3.4.0
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
wcwidth==0.2.14
webdataset==1.0.2
websockets==15.0.1
wheel==0.45.1
